+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+- Improve deleteUnreachable workflow for unreachable clusters (https://github.com/pulumi/pulumi-kubernetes/pull/2489)
+
 ## 3.30.1 (June 29, 2023)
 
 - Add experimental helmChart support to kustomize.Directory (https://github.com/pulumi/pulumi-kubernetes/pull/2471)